Split unit test into header and source
